refactor(cli): split preprocess into csv_input / transform, introduce io module
Reorganize the CLI into pipeline-stage modules so the
read → parse → transform → serialize → write workflow is visible at the
module boundary:

- io.rs: isolates filesystem effects (read_csv, write_json)
- csv_input.rs: CSV row definition, parse_laps_from_csv, and the
  intermediate data types carried downstream
- transform.rs: grouping by car, best-time tracking, position
  calculation
- pipeline.rs: thin orchestration across the stages

The domain types (LapWithMetadata / ExtraData / MiniSectorsRaw) are kept
as-is in this step; deeper type cleanup is deferred to a follow-up.
