+++ cli/cli/src/output.rs
@@ -1,7 +1,7 @@
 use motorsport::{Car, TimelineEvent, calc_time_limit, calc_timeline_events, car, duration};
 use serde::{Serialize, Serializer};
 
-use crate::preprocess::LapWithMetadata;
+use crate::csv_input::LapWithMetadata;
 
 /// セクタータイムのフォーマット
 fn format_sector_time(raw_time: &str, sector_duration: u32) -> String {
