+++ cli/cli/src/pipeline.rs
@@ -1,17 +1,22 @@
-use std::fs;
-use std::path::{Path, PathBuf};
+//! ワークフローのオーケストレーション。
+//!
+//! 1ファイルを処理する際の流れ:
+//!
+//! ```text
+//! read_csv ──▶ parse_laps_from_csv ──▶ group_laps_by_car ──▶ serialize ──▶ write_json
+//! ```
+//!
+//! 各ステージは [`io`] / [`csv_input`] / [`transform`] / [`output`] に分担されており、
+//! このモジュールはそれらを直列に繋ぐだけ。
+
+use std::path::PathBuf;
 
 use crate::FileTask;
+use crate::csv_input::{self, LapWithMetadata};
 use crate::error::CliError;
+use crate::io;
 use crate::output;
-use crate::preprocess;
-
-/// 変換・直列化の結果（純粋な変換の出力）
-struct SerializedOutput {
-    metadata_json: String,
-    laps_json: String,
-    car_count: usize,
-}
+use crate::transform;
 
 /// 処理完了の報告（呼び出し元がログに使う）
 pub struct ProcessingReport {
@@ -20,74 +25,53 @@ pub struct ProcessingReport {
     pub laps_path: PathBuf,
 }
 
-/// パイプライン: read → transform+serialize → write
+/// 1ファイル分のパイプライン: read → parse → transform → serialize → write
 pub fn process_file(task: &FileTask) -> Result<ProcessingReport, CliError> {
-    let csv_content = read_csv(task.input_path())?;
-    let output = transform_and_serialize(&csv_content, task.event_name())?;
-    let written = write_files(task, &output)?;
+    // Stage 1: read
+    let csv_content = io::read_csv(task.input_path())?;
+
+    // Stage 2: parse (CSV → LapWithMetadata のリスト)
+    let laps_with_metadata = csv_input::parse_laps_from_csv(&csv_content);
+
+    // Stage 3: serialize laps & transform (LapWithMetadata → Car のリスト)
+    //
+    // laps は LapWithMetadata から直接構築する。Car には集約済みの情報しか入らないため、
+    // ラップ単体の CSV 由来情報は transform 前に退避する必要がある。
+    let (metadata_json, laps_json, car_count) =
+        serialize(task.event_name(), &laps_with_metadata)?;
+
+    // Stage 4: write
+    let metadata_path = task.output_path().to_path_buf();
+    let laps_path = task.laps_path();
+    io::write_json(&metadata_path, &metadata_json)?;
+    io::write_json(&laps_path, &laps_json)?;
+
     Ok(ProcessingReport {
-        car_count: output.car_count,
-        metadata_path: written.metadata_path,
-        laps_path: written.laps_path,
+        car_count,
+        metadata_path,
+        laps_path,
     })
 }
 
-fn read_csv(input_path: &Path) -> Result<String, CliError> {
-    fs::read_to_string(input_path).map_err(|e| CliError::ReadFile {
-        path: input_path.display().to_string(),
-        source: e,
-    })
-}
-
-fn transform_and_serialize(
-    csv_content: &str,
+/// LapWithMetadata のリストから metadata/laps の両 JSON を組み立てる（副作用なし）。
+fn serialize(
     event_name: &str,
-) -> Result<SerializedOutput, CliError> {
-    let laps_with_metadata = preprocess::parse_laps_from_csv(csv_content);
-    let laps = output::create_laps_output(&laps_with_metadata);
-    let cars = preprocess::group_laps_by_car(laps_with_metadata);
+    laps_with_metadata: &[LapWithMetadata],
+) -> Result<(String, String, usize), CliError> {
+    let laps = output::create_laps_output(laps_with_metadata);
+    let cars = transform::group_laps_by_car(laps_with_metadata.to_vec());
     let car_count = cars.len();
     let metadata = output::create_metadata_output(event_name, &cars);
 
-    let metadata_json = serde_json::to_string_pretty(&metadata).map_err(|e| {
-        CliError::Serialize {
+    let metadata_json =
+        serde_json::to_string_pretty(&metadata).map_err(|e| CliError::Serialize {
             context: "metadata",
             source: e,
-        }
-    })?;
+        })?;
     let laps_json = serde_json::to_string_pretty(&laps).map_err(|e| CliError::Serialize {
         context: "laps",
         source: e,
     })?;
 
-    Ok(SerializedOutput {
-        metadata_json,
-        laps_json,
-        car_count,
-    })
+    Ok((metadata_json, laps_json, car_count))
 }
-
-struct WrittenFiles {
-    metadata_path: PathBuf,
-    laps_path: PathBuf,
-}
-
-fn write_files(task: &FileTask, output: &SerializedOutput) -> Result<WrittenFiles, CliError> {
-    let output_path = task.output_path();
-    let laps_path = task.laps_path();
-
-    fs::write(output_path, &output.metadata_json).map_err(|e| CliError::WriteFile {
-        path: output_path.display().to_string(),
-        source: e,
-    })?;
-    fs::write(&laps_path, &output.laps_json).map_err(|e| CliError::WriteFile {
-        path: laps_path.display().to_string(),
-        source: e,
-    })?;
-
-    Ok(WrittenFiles {
-        metadata_path: output_path.to_path_buf(),
-        laps_path,
-    })
-}
-
